fix messaging
