+++ sbt_common/messaging/src/main/scala/uk/ac/wellcome/messaging/message/MessageReader.scala
@@ -5,10 +5,10 @@ import com.amazonaws.services.sqs.AmazonSQS
 import com.amazonaws.services.sqs
 import uk.ac.wellcome.utils.JsonUtil._
 import com.google.inject.Inject
-import io.circe.Decoder
+import io.circe.{Decoder, Encoder}
 import uk.ac.wellcome.messaging.sns.NotificationMessage
 import uk.ac.wellcome.messaging.sqs.{SQSConfig, SQSReader}
-import uk.ac.wellcome.storage.s3.{S3Config, S3StringStore, S3TypeStore}
+import uk.ac.wellcome.storage.s3.{S3Config, S3TypeStore}
 
 import scala.concurrent.Future
 import uk.ac.wellcome.utils.GlobalExecutionContext.context
@@ -19,7 +19,7 @@ class MessageReader[T] @Inject()(
   messageReaderConfig: MessageReaderConfig,
   s3Client: AmazonS3,
   sqsClient: AmazonSQS
-) {
+)(implicit encoder: Encoder[T], decoder: Decoder[T]) {
   val sqsReader = new SQSReader(sqsClient, messageReaderConfig.sqsConfig)
 
   val s3TypeStore = new S3TypeStore[T](
@@ -28,8 +28,7 @@ class MessageReader[T] @Inject()(
   )
 
   def readAndDelete(process: T => Future[Unit])(
-    implicit decoderN: Decoder[NotificationMessage],
-    decoderT: Decoder[T]
+    implicit decoderN: Decoder[NotificationMessage]
   ): Future[Unit] = {
     sqsReader.retrieveAndDeleteMessages { message =>
       for {
@@ -40,8 +39,7 @@ class MessageReader[T] @Inject()(
   }
 
   private def read(message: sqs.model.Message)(
-    implicit decoderN: Decoder[NotificationMessage],
-    decoderT: Decoder[T]
+    implicit decoderN: Decoder[NotificationMessage]
   ): Future[T] = {
     val deserialisedMessagePointerAttempt = for {
       notification <- fromJson[NotificationMessage](message.getBody)
