+++ sbt_common/messaging/src/main/scala/uk/ac/wellcome/messaging/message/MessageWriter.scala
@@ -4,14 +4,9 @@ import com.amazonaws.services.s3.AmazonS3
 import com.amazonaws.services.sns.AmazonSNS
 import com.google.inject.Inject
 import com.twitter.inject.Logging
-import io.circe.Encoder
+import io.circe.{Decoder, Encoder}
 import uk.ac.wellcome.messaging.sns.{SNSConfig, SNSWriter}
-import uk.ac.wellcome.storage.s3.{
-  KeyPrefixGenerator,
-  S3Config,
-  S3StringStore,
-  S3TypeStore
-}
+import uk.ac.wellcome.storage.s3.{KeyPrefixGenerator, S3Config, S3TypeStore}
 import uk.ac.wellcome.utils.GlobalExecutionContext.context
 import uk.ac.wellcome.utils.JsonUtil._
 
@@ -27,24 +22,19 @@ class MessageWriter[T] @Inject()(
   snsClient: AmazonSNS,
   s3Client: AmazonS3,
   keyPrefixGenerator: KeyPrefixGenerator[T]
-) extends Logging {
+)(implicit encoder: Encoder[T], decoder: Decoder[T]) extends Logging {
 
   private val sns = new SNSWriter(
     snsClient = snsClient,
     snsConfig = messageConfig.snsConfig
   )
 
-  private val s3StringStore = new S3StringStore(
+  private val s3 = new S3TypeStore[T](
     s3Client = s3Client,
     s3Config = messageConfig.s3Config
   )
 
-  private val s3 = new S3TypeStore[T](
-    stringStore = s3StringStore
-  )
-
-  def write(message: T, subject: String)(
-    implicit encoder: Encoder[T]): Future[Unit] = {
+  def write(message: T, subject: String): Future[Unit] = {
     for {
       location <- s3.put(message, keyPrefixGenerator.generate(message))
       pointer <- Future.fromTry(toJson(MessagePointer(location)))
