+++ sbt_common/messaging/src/test/scala/uk/ac/wellcome/messaging/message/MessagingIntegrationTest.scala
@@ -5,7 +5,6 @@ import org.scalatest.{FunSpec, Matchers}
 import uk.ac.wellcome.messaging.test.fixtures.Messaging
 import uk.ac.wellcome.test.fixtures._
 import uk.ac.wellcome.test.utils.ExtendedPatience
-import uk.ac.wellcome.utils.JsonUtil._
 
 class MessagingIntegrationTest
     extends FunSpec
